+++ spec/Pim/Bundle/CatalogBundle/Updater/Copier/MetricAttributeCopierSpec.php
@@ -12,7 +12,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductValue;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 use Prophecy\Argument;
 
-class MetricValueCopierSpec extends ObjectBehavior
+class MetricAttributeCopierSpec extends ObjectBehavior
 {
     function let(
         ProductBuilderInterface $builder,
@@ -42,14 +42,14 @@ class MetricValueCopierSpec extends ObjectBehavior
     ) {
         $fromMetricAttribute->getAttributeType()->willReturn('pim_catalog_metric');
         $toMetricAttribute->getAttributeType()->willReturn('pim_catalog_metric');
-        $this->supports($fromMetricAttribute, $toMetricAttribute)->shouldReturn(true);
+        $this->supportsAttributes($fromMetricAttribute, $toMetricAttribute)->shouldReturn(true);
 
         $fromNumberAttribute->getAttributeType()->willReturn('pim_catalog_number');
         $toNumberAttribute->getAttributeType()->willReturn('pim_catalog_number');
-        $this->supports($fromNumberAttribute, $toNumberAttribute)->shouldReturn(false);
+        $this->supportsAttributes($fromNumberAttribute, $toNumberAttribute)->shouldReturn(false);
 
-        $this->supports($fromMetricAttribute, $toNumberAttribute)->shouldReturn(false);
-        $this->supports($fromNumberAttribute, $toTextareaAttribute)->shouldReturn(false);
+        $this->supportsAttributes($fromMetricAttribute, $toNumberAttribute)->shouldReturn(false);
+        $this->supportsAttributes($fromNumberAttribute, $toTextareaAttribute)->shouldReturn(false);
     }
 
     function it_copies_a_metric_value_to_a_product_value(
@@ -112,7 +112,19 @@ class MetricValueCopierSpec extends ObjectBehavior
         $builder->addProductValue($product3, $toAttribute, $toLocale, $toScope)->shouldBeCalledTimes(1)->willReturn($toProductValue);
 
         $products = [$product1, $product2, $product3, $product4];
-
-        $this->copyValue($products, $fromAttribute, $toAttribute, $fromLocale, $toLocale, $fromScope, $toScope);
+        foreach ($products as $product) {
+            $this->copyAttributeData(
+                $product,
+                $product,
+                $fromAttribute,
+                $toAttribute,
+                [
+                    'from_locale' => $fromLocale,
+                    'to_locale' => $toLocale,
+                    'from_scope' => $fromScope,
+                    'to_scope' => $toScope
+                ]
+            );
+        }
     }
 }