+++ src/Pim/Bundle/CatalogBundle/Updater/Copier/MetricAttributeCopier.php
@@ -3,32 +3,39 @@
 namespace Pim\Bundle\CatalogBundle\Updater\Copier;
 
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilderInterface;
+use Pim\Bundle\CatalogBundle\Factory\MetricFactory;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 
 /**
- * Copy a simple select value attribute in other simple select value attribute
+ * Copy a metric value attribute in other metric value attribute
  *
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class SimpleSelectValueCopier extends AbstractValueCopier
+class MetricAttributeCopier extends AbstractAttributeCopier
 {
+    /** @var MetricFactory */
+    protected $metricFactory;
+
     /**
      * @param ProductBuilderInterface  $productBuilder
      * @param AttributeValidatorHelper $attrValidatorHelper
+     * @param MetricFactory            $metricFactory
      * @param array                    $supportedFromTypes
      * @param array                    $supportedToTypes
      */
     public function __construct(
         ProductBuilderInterface $productBuilder,
         AttributeValidatorHelper $attrValidatorHelper,
+        MetricFactory $metricFactory,
         array $supportedFromTypes,
         array $supportedToTypes
     ) {
         parent::__construct($productBuilder, $attrValidatorHelper);
+        $this->metricFactory  = $metricFactory;
         $this->supportedFromTypes = $supportedFromTypes;
         $this->supportedToTypes = $supportedToTypes;
     }
@@ -51,6 +58,7 @@ class SimpleSelectValueCopier extends AbstractValueCopier
 
         $this->checkLocaleAndScope($fromAttribute, $fromLocale, $fromScope, 'base');
         $this->checkLocaleAndScope($toAttribute, $toLocale, $toScope, 'base');
+        $this->attrValidatorHelper->validateUnitFamilies($fromAttribute, $toAttribute);
 
         $this->copySingleValue(
             $fromProduct,
@@ -69,10 +77,10 @@ class SimpleSelectValueCopier extends AbstractValueCopier
      * @param ProductInterface   $toProduct
      * @param AttributeInterface $fromAttribute
      * @param AttributeInterface $toAttribute
-     * @param string|null        $fromLocale
-     * @param string|null        $toLocale
-     * @param string|null        $fromScope
-     * @param string|null        $toScope
+     * @param string             $fromLocale
+     * @param string             $toLocale
+     * @param string             $fromScope
+     * @param string             $toScope
      */
     protected function copySingleValue(
         ProductInterface $fromProduct,
@@ -86,12 +94,20 @@ class SimpleSelectValueCopier extends AbstractValueCopier
     ) {
         $fromValue = $fromProduct->getValue($fromAttribute->getCode(), $fromLocale, $fromScope);
         if (null !== $fromValue) {
+            $fromData = $fromValue->getData();
             $toValue = $toProduct->getValue($toAttribute->getCode(), $toLocale, $toScope);
             if (null === $toValue) {
                 $toValue = $this->productBuilder->addProductValue($toProduct, $toAttribute, $toLocale, $toScope);
             }
 
-            $toValue->setOption($fromValue->getData());
+            if (null === $metric = $toValue->getMetric()) {
+                $metric = $this->metricFactory->createMetric($fromData->getFamily());
+            }
+
+            $metric->setUnit($fromData->getUnit());
+            $metric->setData($fromData->getData());
+
+            $toValue->setMetric($metric);
         }
     }
 }