+++ src/Pim/Bundle/CatalogBundle/Updater/Copier/MultiSelectAttributeCopier.php
@@ -3,39 +3,33 @@
 namespace Pim\Bundle\CatalogBundle\Updater\Copier;
 
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilderInterface;
-use Pim\Bundle\CatalogBundle\Factory\MetricFactory;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 
 /**
- * Copy a metric value attribute in other metric value attribute
+ * Copy a multi select value attribute in other multi select value attribute
  *
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MetricValueCopier extends AbstractValueCopier
+class MultiSelectAttributeCopier extends AbstractAttributeCopier
 {
-    /** @var MetricFactory */
-    protected $metricFactory;
-
     /**
      * @param ProductBuilderInterface  $productBuilder
      * @param AttributeValidatorHelper $attrValidatorHelper
-     * @param MetricFactory            $metricFactory
      * @param array                    $supportedFromTypes
      * @param array                    $supportedToTypes
      */
     public function __construct(
         ProductBuilderInterface $productBuilder,
         AttributeValidatorHelper $attrValidatorHelper,
-        MetricFactory $metricFactory,
         array $supportedFromTypes,
         array $supportedToTypes
     ) {
         parent::__construct($productBuilder, $attrValidatorHelper);
-        $this->metricFactory  = $metricFactory;
         $this->supportedFromTypes = $supportedFromTypes;
         $this->supportedToTypes = $supportedToTypes;
     }
@@ -58,7 +52,6 @@ class MetricValueCopier extends AbstractValueCopier
 
         $this->checkLocaleAndScope($fromAttribute, $fromLocale, $fromScope, 'base');
         $this->checkLocaleAndScope($toAttribute, $toLocale, $toScope, 'base');
-        $this->attrValidatorHelper->validateUnitFamilies($fromAttribute, $toAttribute);
 
         $this->copySingleValue(
             $fromProduct,
@@ -73,6 +66,8 @@ class MetricValueCopier extends AbstractValueCopier
     }
 
     /**
+     * Copy single value
+     *
      * @param ProductInterface   $fromProduct
      * @param ProductInterface   $toProduct
      * @param AttributeInterface $fromAttribute
@@ -94,20 +89,38 @@ class MetricValueCopier extends AbstractValueCopier
     ) {
         $fromValue = $fromProduct->getValue($fromAttribute->getCode(), $fromLocale, $fromScope);
         if (null !== $fromValue) {
-            $fromData = $fromValue->getData();
             $toValue = $toProduct->getValue($toAttribute->getCode(), $toLocale, $toScope);
             if (null === $toValue) {
                 $toValue = $this->productBuilder->addProductValue($toProduct, $toAttribute, $toLocale, $toScope);
             }
 
-            if (null === $metric = $toValue->getMetric()) {
-                $metric = $this->metricFactory->createMetric($fromData->getFamily());
-            }
+            $this->removeOptions($toValue);
+            $this->copyOptions($fromValue, $toValue);
+        }
+    }
 
-            $metric->setUnit($fromData->getUnit());
-            $metric->setData($fromData->getData());
+    /**
+     * Remove options from attribute
+     *
+     * @param ProductValueInterface $toValue
+     */
+    protected function removeOptions(ProductValueInterface $toValue)
+    {
+        foreach ($toValue->getOptions() as $attributeOption) {
+            $toValue->removeOption($attributeOption);
+        }
+    }
 
-            $toValue->setMetric($metric);
+    /**
+     * Copy attribute options into a multi select attribute
+     *
+     * @param ProductValueInterface $fromValue
+     * @param ProductValueInterface $toValue
+     */
+    protected function copyOptions(ProductValueInterface $fromValue, ProductValueInterface $toValue)
+    {
+        foreach ($fromValue->getOptions() as $attributeOption) {
+            $toValue->addOption($attributeOption);
         }
     }
 }