+++ src/Pim/Bundle/CatalogBundle/Updater/Copier/PriceCollectionAttributeCopier.php
@@ -9,13 +9,13 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 
 /**
- * Copy a multi select value attribute in other multi select value attribute
+ * Copy a price collection value attribute in other price collection value attribute
  *
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MultiSelectValueCopier extends AbstractValueCopier
+class PriceCollectionAttributeCopier extends AbstractAttributeCopier
 {
     /**
      * @param ProductBuilderInterface  $productBuilder
@@ -93,34 +93,21 @@ class MultiSelectValueCopier extends AbstractValueCopier
             if (null === $toValue) {
                 $toValue = $this->productBuilder->addProductValue($toProduct, $toAttribute, $toLocale, $toScope);
             }
-
-            $this->removeOptions($toValue);
             $this->copyOptions($fromValue, $toValue);
         }
     }
 
     /**
-     * Remove options from attribute
-     *
-     * @param ProductValueInterface $toValue
-     */
-    protected function removeOptions(ProductValueInterface $toValue)
-    {
-        foreach ($toValue->getOptions() as $attributeOption) {
-            $toValue->removeOption($attributeOption);
-        }
-    }
-
-    /**
-     * Copy attribute options into a multi select attribute
+     * Copy attribute price into a price collection attribute
      *
      * @param ProductValueInterface $fromValue
      * @param ProductValueInterface $toValue
      */
     protected function copyOptions(ProductValueInterface $fromValue, ProductValueInterface $toValue)
     {
-        foreach ($fromValue->getOptions() as $attributeOption) {
-            $toValue->addOption($attributeOption);
+        foreach ($fromValue->getData() as $price) {
+            $this->productBuilder
+                ->addPriceForCurrencyWithData($toValue, $price->getCurrency(), $price->getData());
         }
     }
 }